split code into two modules for creating widget, and filling widget with data
