+++ index.html
@@ -4,6 +4,7 @@
   <meta charset="utf-8">
   <title>Nashville Software School</title>
   <link rel="stylesheet" href="./styles/styles.css">
+  <link rel="stylesheet" href="./styles/chatWidget_styles.css">
 </head>
 <body>
   <nav></nav>
