+++ scripts/chatWidget/autoScroll.js
@@ -1,0 +1,11 @@
+
+
+// automatically scroll the chatContainer to the bottom when there are more messages than will fit in the window
+
+const autoScroll = function() {
+
+    
+
+}
+
+module.exports = autoScroll
